backend/requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
google-generativeai==0.8.4
python-dotenv==1.0.1
nltk==3.9.1
numpy==2.0.2
scikit-learn==1.4.1.post1
emoji==2.10.1
pytz==2024.1
chromadb==0.4.22
langchain==0.1.9
pandas==2.2.0
spacy==3.7.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
pytest==8.0.1
httpx==0.26.0 